Added IncomeExpense table

+++ SamyuktaManabSoftwares/GrocerySupplyManagementApp/Forms/BankForm.cs
@@ -63,7 +63,7 @@ namespace GrocerySupplyManagementApp.Forms
             MaskEndOfDayFrom.Text = _endOfDay;
             MaskEndOfDayTo.Text = _endOfDay;
             LoadDepositTypes();
-            LoadActionTypes();
+            LoadTransaction();
             EnableFields();
             EnableFields(Action.Load);
         }
@@ -86,23 +86,24 @@ namespace GrocerySupplyManagementApp.Forms
                     {
                         EndOfDay = _endOfDay,
                         BankId = selectedBankId,
-                        Action = ComboActionType.Text.Trim().ToLower() == Constants.DEPOSIT.ToLower() ? '1' : '0',
-                        Debit = ComboActionType.Text.Trim().ToLower() == Constants.DEPOSIT.ToLower() ? Convert.ToDecimal(RichAmount.Text.Trim()) : Constants.DEFAULT_DECIMAL_VALUE,
-                        Credit = ComboActionType.Text.Trim().ToLower() == Constants.DEPOSIT.ToLower() ? Constants.DEFAULT_DECIMAL_VALUE : Convert.ToDecimal(RichAmount.Text.Trim()),
-                        Narration = ComboDepositType.Text.Trim(),
+                        Type = ComboTransaction.Text.Trim().ToLower() == Constants.DEPOSIT.ToLower() ? '1' : '0',
+                        Action = ComboType.Text.Trim(),
+                        Debit = ComboTransaction.Text.Trim().ToLower() == Constants.DEPOSIT.ToLower() ? Convert.ToDecimal(RichAmount.Text.Trim()) : Constants.DEFAULT_DECIMAL_VALUE,
+                        Credit = ComboTransaction.Text.Trim().ToLower() == Constants.DEPOSIT.ToLower() ? Constants.DEFAULT_DECIMAL_VALUE : Convert.ToDecimal(RichAmount.Text.Trim()),
+                        Narration = ComboType.Text.Trim(),
                         AddedBy = _username,
                         AddedDate = DateTime.Now
                     };
 
                     _bankTransactionService.AddBankTransaction(bankTransaction);
-                    DialogResult result = MessageBox.Show(ComboActionType.Text.Trim() + " has been added successfully.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    DialogResult result = MessageBox.Show(ComboTransaction.Text.Trim() + " has been added successfully.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
                     if (result == DialogResult.OK)
                     {
                         var totalBalance = _bankTransactionService.GetTotalBalance(new BankTransactionFilter() { BankId = selectedBankId });
                         TxtBalance.Text = totalBalance.ToString();
-                        ComboActionType.Text = string.Empty;
+                        ComboTransaction.Text = string.Empty;
                         RichAmount.Clear();
-                        ComboDepositType.Text = string.Empty;
+                        ComboType.Text = string.Empty;
                         var bankTransactionViewList = GetBankTransaction();
                         LoadBankTransaction(bankTransactionViewList);
                     }
@@ -134,18 +135,26 @@ namespace GrocerySupplyManagementApp.Forms
                     }
 
                     string selectedId = selectedRow?.Cells["Id"]?.Value?.ToString();
+                    long selectedTransactionId = Convert.ToInt64(selectedRow?.Cells["TransactionId"]?.Value?.ToString());
                     if (!string.IsNullOrWhiteSpace(selectedId))
                     {
-                        DialogResult deleteResult = MessageBox.Show(Constants.MESSAGE_BOX_DELETE_MESSAGE, "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-                        if (deleteResult == DialogResult.Yes)
+                        if(selectedTransactionId == 0)
                         {
-                            var id = Convert.ToInt64(selectedId);
-                            _bankTransactionService.DeleteBankTransaction(id);
-                            var totalBalance = _bankTransactionService.GetTotalBalance(new BankTransactionFilter() { BankId = selectedBankId });
-                            TxtBalance.Text = totalBalance.ToString();
-                            var bankTransactionViewList = GetBankTransaction();
-                            LoadBankTransaction(bankTransactionViewList);
+                            DialogResult deleteResult = MessageBox.Show(Constants.MESSAGE_BOX_DELETE_MESSAGE, "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                            if (deleteResult == DialogResult.Yes)
+                            {
+                                var id = Convert.ToInt64(selectedId);
+                                _bankTransactionService.DeleteBankTransaction(id);
+                                var totalBalance = _bankTransactionService.GetTotalBalance(new BankTransactionFilter() { BankId = selectedBankId });
+                                TxtBalance.Text = totalBalance.ToString();
+                                var bankTransactionViewList = GetBankTransaction();
+                                LoadBankTransaction(bankTransactionViewList);
+                            }
                         }
+                        else
+                        {
+                            DialogResult deleteResult = MessageBox.Show("Please delete main transaction first", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                        } 
                     }
                 }
             }
@@ -262,12 +271,12 @@ namespace GrocerySupplyManagementApp.Forms
 
         private void BtnShowTransaction_Click(object sender, EventArgs e)
         {
-            var action = ComboAction.Text.Trim();
+            var transaction = ComboTransactionFilter.Text.Trim();
             var bankTransactionFilter = new BankTransactionFilter();
 
-            if (!string.IsNullOrWhiteSpace(action))
+            if (!string.IsNullOrWhiteSpace(transaction))
             {
-                bankTransactionFilter.Action = action.Equals(Constants.DEPOSIT) ? '1' : '0';
+                bankTransactionFilter.Type = transaction.Equals(Constants.DEPOSIT) ? '1' : '0';
             }
 
             bankTransactionFilter.DateFrom = UtilityService.GetDate(MaskEndOfDayFrom.Text.Trim());
@@ -290,21 +299,38 @@ namespace GrocerySupplyManagementApp.Forms
         #endregion  
 
         #region Combo Box Event
-        private void ComboDepositType_SelectedValueChanged(object sender, EventArgs e)
+        private void ComboTransaction_SelectedValueChanged(object sender, EventArgs e)
         {
-            ComboActionType.Focus();
+            ComboType.Focus();
         }
 
-        private void ComboActionType_SelectedValueChanged(object sender, EventArgs e)
+        private void ComboType_SelectedValueChanged(object sender, EventArgs e)
         {
             RichAmount.Focus();
         }
+
+        private void ComboTransaction_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            e.Handled = true;
+        }
+
+        private void ComboType_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            e.Handled = true;
+        }
+
+        private void ComboTransactionFilter_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            e.Handled = true;
+        }
+
         #endregion
 
         #region Data Grid Event
         private void DataGridBankDetails_DataBindingComplete(object sender, DataGridViewBindingCompleteEventArgs e)
         {
             DataGridBankList.Columns["Id"].Visible = false;
+            DataGridBankList.Columns["TransactionId"].Visible = false;
 
             DataGridBankList.Columns["EndOfDay"].HeaderText = "Date";
             DataGridBankList.Columns["EndOfDay"].Width = 100;
@@ -373,7 +399,7 @@ namespace GrocerySupplyManagementApp.Forms
                     LoadBankTransaction(bankTransactionViewList);
                     EnableFields();
                     EnableFields(Action.Populate);
-                    ComboActionType.Focus();
+                    ComboTransaction.Focus();
                 }
             }
             catch (Exception ex)
@@ -415,9 +441,9 @@ namespace GrocerySupplyManagementApp.Forms
             }
             else if (action == Action.Populate)
             {
-                ComboActionType.Enabled = true;
+                ComboTransaction.Enabled = true;
                 RichAmount.Enabled = true;
-                ComboDepositType.Enabled = true;
+                ComboType.Enabled = true;
 
                 BtnSaveTransaction.Enabled = true;
                 BtnRemoveTransaction.Enabled = true;
@@ -433,10 +459,10 @@ namespace GrocerySupplyManagementApp.Forms
             {
                 TxtBankName.Enabled = false;
                 TxtAccountNo.Enabled = false;
-                ComboActionType.Enabled = false;
+                ComboTransaction.Enabled = false;
                 RichAmount.Enabled = false;
                 TxtBalance.Enabled = false;
-                ComboDepositType.Enabled = false;
+                ComboType.Enabled = false;
 
                 BtnSaveTransaction.Enabled = false;
                 BtnRemoveTransaction.Enabled = false;
@@ -452,29 +478,29 @@ namespace GrocerySupplyManagementApp.Forms
         {
             TxtBankName.Clear();
             TxtAccountNo.Clear();
-            ComboActionType.Text = string.Empty;
+            ComboTransaction.Text = string.Empty;
             RichAmount.Clear();
             TxtBalance.Clear();
-            ComboDepositType.Text = string.Empty;
+            ComboType.Text = string.Empty;
         }
 
         private void LoadDepositTypes()
         {
-            ComboDepositType.Items.Clear();
-            ComboDepositType.ValueMember = "Id";
-            ComboDepositType.DisplayMember = "Value";
+            ComboType.Items.Clear();
+            ComboType.ValueMember = "Id";
+            ComboType.DisplayMember = "Value";
 
-            ComboDepositType.Items.Add(new ComboBoxItem { Id = Constants.OWNER_EQUITY, Value = Constants.OWNER_EQUITY });
+            ComboType.Items.Add(new ComboBoxItem { Id = Constants.OWNER_EQUITY, Value = Constants.OWNER_EQUITY });
         }
 
-        private void LoadActionTypes()
+        private void LoadTransaction()
         {
-            ComboActionType.Items.Clear();
-            ComboActionType.ValueMember = "Id";
-            ComboActionType.DisplayMember = "Value";
+            ComboTransaction.Items.Clear();
+            ComboTransaction.ValueMember = "Id";
+            ComboTransaction.DisplayMember = "Value";
 
-            ComboActionType.Items.Add(new ComboBoxItem { Id = Constants.DEPOSIT, Value = Constants.DEPOSIT });
-            ComboActionType.Items.Add(new ComboBoxItem { Id = Constants.WITHDRAWL, Value = Constants.WITHDRAWL });
+            ComboTransaction.Items.Add(new ComboBoxItem { Id = Constants.DEPOSIT, Value = Constants.DEPOSIT });
+            ComboTransaction.Items.Add(new ComboBoxItem { Id = Constants.WITHDRAWL, Value = Constants.WITHDRAWL });
         }
         #endregion
 
@@ -505,8 +531,8 @@ namespace GrocerySupplyManagementApp.Forms
         {
             var isValidated = false;
 
-            var depositType = ComboDepositType.Text.Trim();
-            var actionType = ComboActionType.Text.Trim();
+            var depositType = ComboType.Text.Trim();
+            var actionType = ComboTransaction.Text.Trim();
             var amount = RichAmount.Text.Trim();
 
             if (string.IsNullOrWhiteSpace(depositType)
@@ -526,5 +552,6 @@ namespace GrocerySupplyManagementApp.Forms
             return isValidated;
         }
         #endregion
+
     }
 }
